Enhanced

+++ internal/modules/services/topology/dto.go
@@ -5,7 +5,7 @@ type nodeAggRow struct {
 	ServiceName  string    `ch:"service"`
 	RequestCount uint64    `ch:"request_count"`
 	ErrorCount   uint64    `ch:"error_count"`
-	QS           []float32 `ch:"qs"`
+	QS           []float64 `ch:"qs"`
 	P50Ms        float32   `ch:"p50_ms"`
 	P95Ms        float32   `ch:"p95_ms"`
 	P99Ms        float32   `ch:"p99_ms"`
@@ -17,7 +17,7 @@ type edgeAggRow struct {
 	Target     string    `ch:"target"`
 	CallCount  uint64    `ch:"call_count"`
 	ErrorCount uint64    `ch:"error_count"`
-	QS         []float32 `ch:"qs"`
+	QS         []float64 `ch:"qs"`
 	P50Ms      float32   `ch:"p50_ms"`
 	P95Ms      float32   `ch:"p95_ms"`
 }
